requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.12
pydantic-settings==2.10.1
python-dotenv==1.1.1
boto3==1.40.1
python-multipart==0.0.20
httpx==0.28.1
pypdf==6.7.5
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4

bcrypt==4.0.1
Pillow==11.2.1

language-tool-python==2.9.4
openpyxl==3.1.5
